refactor tests for post update triggers
